pyproject.toml
[project]
name = "professor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.68.0",
    "colorama>=0.4.6",
    "cutie>=0.3.2",
    "docker>=7.1.0",
    "jq>=1.10.0",
    "python-dotenv>=1.1.1",
]
